constant time hash
